# cmd/docs/openapiv3.json
{"components":{"requestBodies":{"getPendingCSRRequest":{"content":{"application/json":{"schema":{"properties":{"ID":{"type":"integer"}}}}},"description":"Request used for creating a new CSR Form","required":true},"postCSRFormRequest":{"content":{"application/json":{"schema":{"properties":{"subject":{"$ref":"#/components/schemas/CSRForm"}}}}},"description":"Request used for creating a new CSR Form","required":true},"postCSRRequest":{"content":{"application/json":{"schema":{"properties":{"data":{"type":"string"},"dms_name":{"type":"string"},"key_type":{"type":"string"}}}}},"description":"Request used for creating a new CSR ","required":true}},"responses":{"DeleteCSRResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after deleting CSR."},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"}}}}},"description":"Response when errors happen."},"GetCRTResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after deleting CSR."},"GetPendingCSRDBResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CSR"}}},"description":"Response returned back after getting pending CSRDB."},"GetPendingCSRFileResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after getting CSR File."},"GetPendingCSRsResponse":{"content":{"application/json":{"schema":{"properties":{"CSRs":{"items":{"$ref":"#/components/schemas/CSR"},"type":"array"}}}}},"description":"Response returned back after getting pending CSRs."},"HealthResponse":{"content":{"application/json":{"schema":{"properties":{"healthy":{"type":"boolean"}}}}},"description":"Response returned back after healthchecking."},"PostCSRFormResponse":{"content":{"application/json":{"schema":{"properties":{"csr":{"$ref":"#/components/schemas/CSR"},"priv_key":{"type":"string"}}}}},"description":"Response returned back after creating a CSR."},"PostCSRResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CSR"}}},"description":"Response returned back after getting a CA."},"PutChangeCSRStatusResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CSR"}}},"description":"Response returned back after changing CSR Status."}},"schemas":{"CSR":{"properties":{"common_name":{"type":"string"},"country":{"type":"string"},"csr":{"type":"string"},"dms_name":{"type":"string"},"id":{"type":"integer"},"locality":{"type":"string"},"mail":{"type":"string"},"organization":{"type":"string"},"organization_unit":{"type":"string"},"state":{"type":"string"},"status":{"type":"string"},"url":{"type":"string"}},"type":"object"},"CSRForm":{"properties":{"common_name":{"type":"string"},"country":{"type":"string"},"dms_name":{"type":"string"},"key_bits":{"type":"integer"},"key_type":{"type":"string"},"locality":{"type":"string"},"mail":{"type":"string"},"organization":{"type":"string"},"organization_unit":{"type":"string"},"state":{"type":"string"},"url":{"type":"string"}},"type":"object"}}},"info":{"contact":{"url":"https://github.com/lamassuiot"},"description":"REST API used for interacting with Lamassu DMS Enroller","license":{"name":"MPL v2.0","url":"https://github.com/lamassuiot/lamassu-compose/blob/main/LICENSE"},"title":"Lamassu DMS Enroller API","version":"0.0.0"},"openapi":"3.0.0","paths":{"/v1/csrs":{"get":{"description":"Get Pending CSRs","operationId":"GetPendingCSRs","responses":{"200":{"$ref":"#/components/responses/GetPendingCSRsResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/csrs/{id}":{"delete":{"description":"Delete CSR by id","operationId":"DeleteCSR","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/DeleteCSRResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"get":{"description":"Get Pending CSRDB by id","operationId":"GetPendingCSRDB","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/GetPendingCSRDBResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"put":{"description":"Change CSR Status by id","operationId":"PutChangeCSRStatus","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/PutChangeCSRStatusResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/csrs/{id}/crt":{"get":{"description":"Get CRT by id","operationId":"GetCRT","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/GetCRTResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/csrs/{id}/file":{"get":{"description":"Get Pending CSR Fileby id","operationId":"GetPendingCSRFile","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/GetPendingCSRFileResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/csrs/{name}":{"post":{"description":"Post CSR","operationId":"PostCSR","parameters":[{"in":"path","name":"name","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/postCSRRequest"},"responses":{"200":{"$ref":"#/components/responses/PostCSRResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/csrs/{name}/form":{"post":{"description":"Post CSR Form","operationId":"PostCSRForm","parameters":[{"in":"path","name":"name","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/postCSRFormRequest"},"responses":{"200":{"$ref":"#/components/responses/PostCSRFormResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/health":{"get":{"description":"Get health status","operationId":"Health","responses":{"200":{"$ref":"#/components/responses/HealthResponse"}}}}},"servers":[{"description":"Current Server","url":"/"}]}